--- server/exerciseAPIFunctions.js
@@ -1,9 +1,0 @@
-const express = require('express')
-const app = express()
-const port = process.env.PORT || 3000
-
-module.exports = {
-   users: function() {
-      return "This is community users";
-   }
-}
